# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a IM System API.",
        "title": "IM System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/friend-requests/incoming": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回当前用户收到的待处理好友申请列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "好友申请"
                ],
                "summary": "获取收到的好友申请",
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.FriendRequestListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/friend-requests/outgoing": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回当前用户发出的待处理好友申请列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "好友申请"
                ],
                "summary": "获取发出的好友申请",
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.FriendRequestListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/friend-requests/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "向指定用户发送好友申请；若存在反向待处理申请，则自动同意",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "好友申请"
                ],
                "summary": "发送好友申请",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "附言",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.SendFriendRequestReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "发送成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "目标用户不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "409": {
                        "description": "申请状态冲突",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/friend-requests/{id}/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "同意指定的好友申请，并建立好友关系与单聊会话",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "好友申请"
                ],
                "summary": "同意好友申请",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "好友申请ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "处理成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "好友申请不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "409": {
                        "description": "好友申请状态不合法",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/friend-requests/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "拒绝指定的好友申请",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "好友申请"
                ],
                "summary": "拒绝好友申请",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "好友申请ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "处理成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "好友申请不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "409": {
                        "description": "好友申请状态不合法",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.FriendRequestItemResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "receiver": {
                    "$ref": "#/definitions/dto.FriendRequestUserResp"
                },
                "requester": {
                    "$ref": "#/definitions/dto.FriendRequestUserResp"
                },
                "status": {
                    "$ref": "#/definitions/model.FriendRequestStatus"
                }
            }
        },
        "dto.FriendRequestListResp": {
            "type": "object",
            "properties": {
                "requests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.FriendRequestItemResp"
                    }
                }
            }
        },
        "dto.FriendRequestUserResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "online": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "dto.SendFriendRequestReq": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.FriendRequestStatus": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "FriendRequestPending",
                "FriendRequestAccepted",
                "FriendRequestRejected"
            ]
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}